requirements-test.txt
wheel==0.37.0
responses==0.13.4
numpy==1.22.0
Pillow==9.0.1
Django==3.2.20
django-bulk-update==2.2.0
django-admin-tools==0.9.0
django-recaptcha==2.0.6
django-markdown2==0.3.1
django-comments-xtd==2.9.2
django-extensions==3.1.3
django-hosts==4.0
django-autocomplete-light==3.8.2
sorl-thumbnail==12.11.0
django-compressor==2.4.1
django-modeltranslation==0.17.3
django-haystack==3.0
django-registration-redux==2.9
django-bootstrap4==3.0.1
django-leaflet==0.28.1
django-user-agents==0.4.0
django-allauth==0.55.0
django-cors-headers
flickrapi==2.4.0
geopy==2.2.0
google-auth==2.0.0
opencv-python==4.5.3.56
pandas==1.1.5
psycopg-binary==3.3.2
requests==2.26.0
roman==3.3
sklearn
oauth2client==4.1.3
gdal==3.0.4
typing==3.7.4.3
mypy==0.910
pytest==6.2.4
pytest-cov==2.12.1
pytest-django==4.4.0
pysolr==3.9.0
django-stubs==1.8.0
djangorestframework-stubs==1.4.0
pytest-mypy==0.8.1
pytest-mypy-plugins==1.7.0
flake8==3.9.2
